Followed your feedback and made the changes.
I have a question:
For the  @Test void getSecondMessageFromList(), the output is not correct, I am trying to assert the value from my key=phoneNumber, from my map messageList, to the expected use of the method getSecondMessage.
I receive the memory address and did not figure out how to write @toString, or what to use?
Thx :)